refactor: move functions from helpers to module where they are used
