+++ crates/pop-cli/src/commands/new/parachain.rs
@@ -1,17 +1,14 @@
 // SPDX-License-Identifier: GPL-3.0
-use crate::{
-	helpers::{
-		check_destination_path, display_release_versions_to_user, display_select_options,
-		get_customization_value, is_template_supported, prompt_customizable_options,
-	},
-	style::{style, Theme},
-};
+use crate::style::{style, Theme};
 use anyhow::Result;
 use clap::Args;
-use std::path::PathBuf;
+use std::{
+	fs,
+	path::{Path, PathBuf},
+};
 
-use cliclack::{clear_screen, input, intro, log, outro, outro_cancel, set_theme};
-use pop_parachains::{instantiate_template_dir, Config, Git, GitHub, Provider, Template};
+use cliclack::{clear_screen, confirm, input, intro, log, outro, outro_cancel, set_theme};
+use pop_parachains::{instantiate_template_dir, Config, Git, GitHub, Provider, Release, Template};
 
 #[derive(Args)]
 pub struct NewParachainCommand {
@@ -73,44 +70,6 @@ impl NewParachainCommand {
 	}
 }
 
-fn generate_parachain_from_template(
-	name_template: &String,
-	provider: &Provider,
-	template: &Template,
-	tag_version: Option<String>,
-	config: Config,
-) -> Result<()> {
-	intro(format!(
-		"{}: Generating \"{}\" using {:?} from {:?}!",
-		style(" Pop CLI ").black().on_magenta(),
-		name_template,
-		template,
-		provider
-	))?;
-	let destination_path = check_destination_path(name_template)?;
-
-	let mut spinner = cliclack::spinner();
-	spinner.start("Generating parachain...");
-	let tag = instantiate_template_dir(template, destination_path, tag_version, config)?;
-	if let Err(err) = Git::git_init(destination_path, "initialized parachain") {
-		if err.class() == git2::ErrorClass::Config && err.code() == git2::ErrorCode::NotFound {
-			outro_cancel("git signature could not be found. Please configure your git config with your name and email")?;
-		}
-	}
-	spinner.stop("Generation complete");
-	if let Some(tag) = tag {
-		log::info(format!("Version: {}", tag))?;
-	}
-
-	cliclack::note(
-			"NOTE: the resulting parachain is not guaranteed to be audited or reviewed for security vulnerabilities.",
-		format!("Please consult the source repository at {} to assess production suitability and licensing restrictions.", template.repository_url()?))?;
-
-	outro(format!("cd into \"{}\" and enjoy hacking! 🚀", name_template))?;
-
-	Ok(())
-}
-
 async fn guide_user_to_generate_parachain() -> Result<()> {
 	intro(format!("{}: Generate a parachain", style(" Pop CLI ").black().on_magenta()))?;
 
@@ -166,6 +125,143 @@ async fn guide_user_to_generate_parachain() -> Result<()> {
 	)
 }
 
+fn generate_parachain_from_template(
+	name_template: &String,
+	provider: &Provider,
+	template: &Template,
+	tag_version: Option<String>,
+	config: Config,
+) -> Result<()> {
+	intro(format!(
+		"{}: Generating \"{}\" using {:?} from {:?}!",
+		style(" Pop CLI ").black().on_magenta(),
+		name_template,
+		template,
+		provider
+	))?;
+	let destination_path = check_destination_path(name_template)?;
+
+	let mut spinner = cliclack::spinner();
+	spinner.start("Generating parachain...");
+	let tag = instantiate_template_dir(template, destination_path, tag_version, config)?;
+	if let Err(err) = Git::git_init(destination_path, "initialized parachain") {
+		if err.class() == git2::ErrorClass::Config && err.code() == git2::ErrorCode::NotFound {
+			outro_cancel("git signature could not be found. Please configure your git config with your name and email")?;
+		}
+	}
+	spinner.stop("Generation complete");
+	if let Some(tag) = tag {
+		log::info(format!("Version: {}", tag))?;
+	}
+
+	cliclack::note(
+			"NOTE: the resulting parachain is not guaranteed to be audited or reviewed for security vulnerabilities.",
+		format!("Please consult the source repository at {} to assess production suitability and licensing restrictions.", template.repository_url()?))?;
+
+	outro(format!("cd into \"{}\" and enjoy hacking! 🚀", name_template))?;
+
+	Ok(())
+}
+
+fn is_template_supported(provider: &Provider, template: &Template) -> Result<()> {
+	if !template.matches(provider) {
+		return Err(anyhow::anyhow!(format!(
+			"The provider \"{:?}\" doesn't support the {:?} template.",
+			provider, template
+		)));
+	};
+	return Ok(());
+}
+
+fn display_select_options(provider: &Provider) -> Result<&Template> {
+	let mut prompt = cliclack::select("Select the type of parachain:".to_string());
+	for (i, template) in provider.templates().into_iter().enumerate() {
+		if i == 0 {
+			prompt = prompt.initial_value(template);
+		}
+		prompt = prompt.item(template, template.name(), template.description());
+	}
+	Ok(prompt.interact()?)
+}
+
+fn get_customization_value(
+	template: &Template,
+	symbol: Option<String>,
+	decimals: Option<u8>,
+	initial_endowment: Option<String>,
+) -> Result<Config> {
+	if !matches!(template, Template::Base)
+		&& (symbol.is_some() || decimals.is_some() || initial_endowment.is_some())
+	{
+		log::warning("Customization options are not available for this template")?;
+	}
+	return Ok(Config {
+		symbol: symbol.clone().expect("default values"),
+		decimals: decimals.clone().expect("default values"),
+		initial_endowment: initial_endowment.clone().expect("default values"),
+	});
+}
+
+fn check_destination_path(name_template: &String) -> Result<&Path> {
+	let destination_path = Path::new(name_template);
+	if destination_path.exists() {
+		if !confirm(format!(
+			"\"{}\" directory already exists. Would you like to remove it?",
+			destination_path.display()
+		))
+		.interact()?
+		{
+			outro_cancel(format!(
+				"Cannot generate parachain until \"{}\" directory is removed.",
+				destination_path.display()
+			))?;
+			return Err(anyhow::anyhow!(format!(
+				"\"{}\" directory already exists.",
+				destination_path.display()
+			)));
+		}
+		fs::remove_dir_all(destination_path)?;
+	}
+	Ok(destination_path)
+}
+
+fn display_release_versions_to_user(releases: Vec<Release>) -> Result<String> {
+	let mut prompt = cliclack::select("Select a specific release:".to_string());
+	for (i, release) in releases.iter().enumerate() {
+		if i == 0 {
+			prompt = prompt.initial_value(&release.tag_name);
+		}
+		prompt = prompt.item(
+			&release.tag_name,
+			&release.name,
+			match &release.commit {
+				Some(commit) => format!("{} / {}", &release.tag_name, &commit[..=6]),
+				None => release.tag_name.to_string(),
+			},
+		)
+	}
+	Ok(prompt.interact()?.to_string())
+}
+
+fn prompt_customizable_options() -> Result<Config> {
+	let symbol: String = input("What is the symbol of your parachain token?")
+		.placeholder("UNIT")
+		.default_input("UNIT")
+		.interact()?;
+
+	let decimals_input: String = input("How many token decimals?")
+		.placeholder("12")
+		.default_input("12")
+		.interact()?;
+	let decimals: u8 = decimals_input.parse::<u8>().expect("input has to be a number");
+
+	let endowment: String = input("And the initial endowment for dev accounts?")
+		.placeholder("1u64 << 60")
+		.default_input("1u64 << 60")
+		.interact()?;
+	Ok(Config { symbol, decimals, initial_endowment: endowment })
+}
+
 #[cfg(test)]
 mod tests {
 
@@ -198,5 +294,35 @@ mod tests {
 			eprintln!("Failed to delete directory: {}", err);
 		}
 		Ok(())
+	}
+
+	#[test]
+	fn test_is_template_supported() {
+		assert!(is_template_supported(&Provider::Pop, &Template::Base).is_ok());
+		assert!(is_template_supported(&Provider::Pop, &Template::ParityContracts).is_err());
+		assert!(is_template_supported(&Provider::Pop, &Template::ParityFPT).is_err());
+
+		assert!(is_template_supported(&Provider::Parity, &Template::Base).is_err());
+		assert!(is_template_supported(&Provider::Parity, &Template::ParityContracts).is_ok());
+		assert!(is_template_supported(&Provider::Parity, &Template::ParityFPT).is_ok());
+	}
+
+	#[test]
+	fn test_get_customization_values() {
+		let config = get_customization_value(
+			&Template::Base,
+			Some("DOT".to_string()),
+			Some(6),
+			Some("10000".to_string()),
+		);
+		assert!(config.is_ok());
+		assert_eq!(
+			config.unwrap(),
+			Config {
+				symbol: "DOT".to_string(),
+				decimals: 6,
+				initial_endowment: "10000".to_string()
+			}
+		);
 	}
 }
